app-backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Configuration and environment
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# LangChain and AI (compatible versions)
langchain-core==0.3.15
langchain==0.3.7
langgraph==0.2.45
langgraph-checkpoint-postgres==2.0.6
psycopg[pool]==3.2.3
# Note: Using older langgraph for now to maintain compatibility
# Checkpointing works with this version using PostgresSaver from langgraph.checkpoint.postgres

# Vector Database and Embeddings
qdrant-client==1.16.1
sentence-transformers==5.1.2

# Scientific Computing (required by sentence-transformers and A/B testing)
scipy==1.11.4
numpy==1.26.2

# HTTP clients
requests==2.32.3
httpx==0.27.2

# Background Jobs and Task Queue
celery==5.4.0
redis==5.0.1

# LLM Provider SDKs
google-generativeai==0.8.3

# Security and scanning tools
python-nmap==0.7.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Logging
loguru==0.7.2

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Development (optional)
ruff==0.8.4
mypy==1.13.0
bandit==1.8.0
pre-commit==4.0.1

# Agentic RAG (Phase 3)
networkx==3.6
cpe==1.2.1
nest_asyncio==1.6.0

# Optimization (Phase 4.2)
tiktoken==0.5.2  # Accurate token counting for LLMs
locust==2.20.0  # Load testing framework
